Jpa 와 Auth 통합

+++ src/main/kotlin/com/bifos/piudbbackend/service/AppUserService.kt
@@ -10,6 +10,6 @@ interface AppUserService {
 
     fun findUsersByEmail(partialEmail: String): List<AppUser>
 
-    fun createUser(user: AppUser): Int
+    fun createUser(user: AppUser): Long
 
 }
